Removed unneeded include
We can't write unit tests for concrete methods of abstract class,
because we can't create instance to call method
We could have added include for subclass, but this would result in
circular include
We could also define subclass just for test purposes - but this is
probably not needed, because there are already tests for this method in
subclass implementation.
So we just add comment that explains where to find these tests.