Trying to fix warning with char* literals
We have declared CommandLineProcessor::run to accept same parameters as
main(int argc, char** argv)
In fact, it doesn't need to modify strings, so it would work with const
char** as well
But in this case, we must modify main, or we won't be able to pass argv
(defined as char**) to a function expecting const char**

Using main with signature int main(int argc, const char** argv) seems to
work, and it probably results in exactly same binary code as without
const.
However, C++ standard doesn't require compiler to support this version
of main - so this code may not work in some compilers.